choozenREST/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
mysqlclient==2.0.1
django-mysql==3.9
django-cors-headers==3.9.0
requests==2.27.0
